llama_api.py
# Placeholder for LLaMA API interaction
# We will need a library like 'requests' or a specific Llama client
import os
from dotenv import load_dotenv

load_dotenv() # Load environment variables from .env file

# Replace with your actual Llama API key and endpoint if needed
# LLAMA_API_KEY = os.getenv("LLAMA_API_KEY")
# LLAMA_API_ENDPOINT = os.getenv("LLAMA_ENDPOINT", "DEFAULT_LLAMA_ENDPOINT_HERE")

def call_llama_api(prompt, images=None):
    """Placeholder function to simulate calling the Llama API."""
    print("--- Calling Llama API (Simulated) ---")
    print(f"Prompt: {prompt}")
    if images:
        print(f"Images: {[img[:10] + '...' if len(img) > 10 else img for img in images]}") # Print truncated image names/paths

    # Simulate a response based on the type of request
    if "provide a short, evocative clue" in prompt:
        response = {"clue": "Ephemeral Dreams", "chosen_card_index": 0} # Dummy clue and choice
    elif "choose the card that best matches the clue" in prompt:
        response = {"chosen_card_index": 1} # Dummy guess
    else:
        response = {"error": "Unknown prompt type"}

    print(f"Response: {response}")
    print("--------------------------------------")
    return response

def generate_prompt_for_clue(hand_images):
    """Generates a prompt for the Llama model to provide a clue."""
    # TODO: Refine prompt engineering
    prompt = f"You are playing Dixit. Your hand consists of the following cards (images provided separately): {', '.join(hand_images)}. Provide a short, evocative clue for one of your cards and indicate which card you chose (by its index in the list)."
    return prompt

def generate_prompt_for_guess(clue, displayed_images):
    """Generates a prompt for the Llama model to guess the storyteller's card."""
    # TODO: Refine prompt engineering
    prompt = f"You are playing Dixit. The storyteller's clue is '{clue}'. The cards on the table are (images provided separately): {', '.join(displayed_images)}. Choose the card that best matches the clue (by its index in the list)."
    return prompt

def get_clue_from_llama(hand_images):
    """Gets a clue and chosen card index from Llama."""
    prompt = generate_prompt_for_clue(hand_images)
    # In a real scenario, you'd handle image data appropriately
    response = call_llama_api(prompt, images=hand_images)
    # TODO: Add error handling and proper response parsing
    clue = response.get("clue", "Default Clue")
    chosen_card_index = response.get("chosen_card_index", 0)
    # We need the actual card, assuming hand_images is ordered
    if 0 <= chosen_card_index < len(hand_images):
        chosen_card = hand_images[chosen_card_index]
        # Important: Remove the chosen card from the hand representation if needed outside this function
        return clue, chosen_card
    else:
        print("Error: Llama returned invalid card index for clue.")
        return "Error Clue", hand_images[0] # Fallback

def get_guess_from_llama(clue, displayed_images):
    """Gets a card guess index from Llama."""
    prompt = generate_prompt_for_guess(clue, displayed_images)
    # In a real scenario, you'd handle image data appropriately
    response = call_llama_api(prompt, images=displayed_images)
    # TODO: Add error handling and proper response parsing
    chosen_card_index = response.get("chosen_card_index", 0)
    if 0 <= chosen_card_index < len(displayed_images):
         return chosen_card_index
    else:
         print("Error: Llama returned invalid card index for guess.")
         return 0 # Fallback 